Write educator class, people class and teacher module, pass in inherticance. Answer markdown questions all rspec pass
